Eliminado Disponible y sustituido por la clase Huesped. Modificada funcionalidad del clineteweb para Huesped. Tanto buscar, disponibles y no disponibles, funcionan correctamente. Modificaciones del estado de Disponible a No disponible funcionando

+++ gestion-apartamentos/src/main/java/com/saguadopro/gestionapartamentos/repositories/HuespedRepo.java
@@ -3,6 +3,6 @@ package com.saguadopro.gestionapartamentos.repositories;
 import com.saguadopro.gestionapartamentos.entities.Huesped;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface HuespedRepo extends JpaRepository<Huesped,Integer> {
+public interface HuespedRepo extends JpaRepository<Huesped,Long> {
 
 }
